feat: hooks 함수를 사용하여 페이지 분리
